+++ Automation/ScriptingEngine/ScriptableComponents/Components/Configurator.cs
@@ -28,6 +28,7 @@ sealed class Configurator : IConfigurator {
     containerDefinition.Bind<InventoryScriptableComponent>().AsSingleton();
     containerDefinition.Bind<InventoryScriptableComponent.EmptyingStatusBehavior>().AsTransient();
     containerDefinition.Bind<InventoryScriptableComponent.InventoryChangeTracker>().AsTransient();
+    containerDefinition.Bind<LeverScriptableComponent>().AsSingleton();
     containerDefinition.Bind<ManufactoryScriptableComponent>().AsSingleton();
     containerDefinition.Bind<PausableScriptableComponent>().AsSingleton();
     containerDefinition.Bind<PlantableScriptableComponent>().AsSingleton();